+++ ydb/core/blob_depot/agent/ya.make
@@ -1,5 +1,15 @@
 LIBRARY()
 
+    IF (OS_WINDOWS)
+        CFLAGS(
+            -DKIKIMR_DISABLE_S3_OPS
+        )
+    ELSE()
+        SRCS(
+            s3.cpp
+        )
+    ENDIF()
+
     SRCS(
         agent.cpp
         agent.h
