+++ ydb/core/blob_depot/s3_delete.cpp
@@ -1,8 +1,11 @@
 #include "s3.h"
 
+#include <ydb/core/wrappers/abstract.h>
+
 namespace NKikimr::NBlobDepot {
 
     using TS3Manager = TBlobDepot::TS3Manager;
+    using TEvExternalStorage = NWrappers::TEvExternalStorage;
 
     struct TS3Manager::TEvDeleteResult : TEventLocal<TEvDeleteResult, TEvPrivate::EvDeleteResult> {
         std::vector<TS3Locator> LocatorsOk;
