+++ ydb/core/blob_depot/ya.make
@@ -1,5 +1,22 @@
 LIBRARY()
 
+    IF (OS_WINDOWS)
+        CFLAGS(
+            -DKIKIMR_DISABLE_S3_OPS
+        )
+        SRCS(
+            s3_windows_stub.cpp
+        )
+    ELSE()
+        SRCS(
+            s3.cpp
+            s3_delete.cpp
+            s3_scan.cpp
+            s3_upload.cpp
+            s3_write.cpp
+        )
+    ENDIF()
+
     SRCS(
         blob_depot.cpp
         blob_depot.h
@@ -33,11 +50,6 @@ LIBRARY()
         group_metrics_exchange.cpp
         mon_main.cpp
         mon_main.h
-        s3.cpp
-        s3_delete.cpp
-        s3_scan.cpp
-        s3_upload.cpp
-        s3_write.cpp
         s3.h
         space_monitor.cpp
         space_monitor.h
