+++ ydb/core/util/aws_windows_stub.cpp
@@ -1,0 +1,8 @@
+#include "aws.h"
+
+namespace NKikimr {
+
+void InitAwsAPI() {}
+void ShutdownAwsAPI() {}
+
+} // NKikimr
